refactor(ai-provider): make the AI SDK path private to AiService
Every module in provider/ was checked against its consumers: provider
config (686 lines), Vertex authentication, and model listing are reached
only from the AI SDK path, so they move into generation/ alongside the
generator they serve. What stays in provider/ is what both execution
paths share — connection facts and the system model support factory —
which is what the layer claims to be.

generation/ also loses two redundant directory levels: aiSdk/ had no
sibling implementation to distinguish, and params/ held one file. A lint
rule now keeps generation/ private to AiService.ts so the facade stays
the only entry point.
