+++ src/backend/ai/README.md
@@ -7,8 +7,9 @@ owns the mobile platform and application-service boundaries around that package.
 
 - `AiService.ts` owns non-conversation generation, model listing, model checks, and image generation.
   Callers supply an explicit `uniqueModelId`; it does not resolve Agent state or stream chat turns.
-- `generation/` owns non-conversation AI SDK request assembly, execution, and mobile usage capture
-  used by `AiService`. It owns no persisted turn state.
+- `generation/` is the private implementation of `AiService`: AI SDK request assembly, provider
+  configuration, model listing, Vertex authentication, execution, and mobile usage capture. Nothing
+  outside `AiService.ts` imports it (ESLint-enforced). It owns no persisted turn state.
 - `agent/runtime/` owns the independent Agent Runtime contract, its FakeRuntime test double, and the
   Pi Runtime implementation (`docs/references/agent/agent-runtime.md`). This boundary must not
   import application protocol types, persistence, React, or Expo modules (ESLint-enforced).
@@ -19,11 +20,11 @@ owns the mobile platform and application-service boundaries around that package.
 - `agent/sessionStore/`, `agent/resources/`, `agent/tools/`, and `agent/piAdapter/` respectively own
   transcript persistence, managed turn resources, executable Agent capabilities, and the mobile
   provider/model adaptation required to construct Pi.
-- `provider/` resolves credential-selection-free connection facts shared by Pi and AI SDK request
-  construction, owns shared Provider HTTP transport policies and the runtime-agnostic system model
-  support factory, injects Expo environment values and app headers, then builds AI SDK provider
-  configuration from mobile data services. The Pi language compatibility decision lives in
-  `agent/piAdapter/`, behind the Runtime binding.
+- `provider/` is the runtime-agnostic connection-fact layer, and nothing more: it resolves
+  credential-selection-free connection facts shared by Pi and AI SDK request construction, and
+  declares the system model support factory whose language answer comes from the bound Runtime.
+  Anything used by only one execution path belongs to that path — AI SDK provider configuration
+  lives in `generation/`, and the Pi language compatibility decision lives in `agent/piAdapter/`.
 - `mcp/` owns the mobile Streamable HTTP transport, connection lifecycle, server status, and tool
   discovery used by MCP settings.
 - `devBench/` owns development-only provider fixtures and benchmark request helpers.
