+++ src/backend/ai/__tests__/AiService.test.ts
@@ -12,7 +12,7 @@ jest.mock('@cherrystudio/ai-core', () => ({
   generateImage: jest.fn(),
 }));
 
-jest.mock('@/backend/ai/generation/aiSdk/AiSdkGenerator', () => ({
+jest.mock('@/backend/ai/generation/AiSdkGenerator', () => ({
   AiSdkGenerator: jest.fn().mockImplementation((params) => {
     mockGeneratorConstructor(params);
     return { generate: mockGenerate };
