+++ src/backend/ai/generation/AiSdkGenerator.ts
