+++ src/backend/ai/generation/VertexAuthClient.ts
