+++ src/backend/ai/generation/__tests__/AiSdkGenerator.test.ts
