+++ src/backend/ai/generation/__tests__/VertexAuthClient.test.ts
