+++ src/backend/ai/generation/__tests__/buildAgentParams.test.ts
