+++ src/backend/ai/generation/__tests__/reasoningCatalog.test.ts
