--- src/backend/ai/generation/aiSdk/index.ts
@@ -1,2 +1,0 @@
-export { AiSdkGenerator } from './AiSdkGenerator';
-export { buildAgentParams } from './params/buildAgentParams';
