+++ src/backend/ai/generation/index.ts
@@ -1,0 +1,2 @@
+export { AiSdkGenerator } from './AiSdkGenerator';
+export { buildAgentParams } from './buildAgentParams';
